submit button copy fix & refactor styles
